requirements.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
openpyxl
numpy
pandas
easyocr